requirements.txt
biopython==1.85
distributed==2025.1.0
dask==2025.1.0
google-cloud-storage==2.19.0
matplotlib==3.10.0
pandas==2.2.3
numpy==2.2.2
tqdm==4.67.1
